Create BorrowList and BorrowDetailsDialog
